trying currentLocation is bigInt for postgres

+++ data/migrations/20200819113647_create-trucks-table.js
@@ -6,7 +6,7 @@ exports.up = function (knex) {
     tbl.string('cuisineType', 128).notNullable();
     tbl.string('currentLocation', 256).notNullable();
     tbl
-      .integer('departureTime')
+      .bigInteger('departureTime')
       .unsigned()
       .notNullable()
       .defaultTo(new Date().getTime());
